Update admin menu template.
